pyproject.toml
[tool.poetry]
name = "lol-app"
version = "0.1.0"
description = "League of Legends complex app"
authors = ["Alexandr Obolonkov <itbolonkov2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = {extras = ["voice"], version = "^2.3.2"}


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["F", "E", "W", "C90",
    "I", "N", "S", "B", "A",
    "ISC", "T20", "Q", "PTH"]
ignore = ["T201"]

[tool.mypy]
explicit_package_bases = true